+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added game edition switcher
 - Added changelog to updated components toast
 - Added wine tools to settings
+- Added preferences search
 
 ### Fixed